Remove hard dependency on `lsp-mode`
This removes the hard dependency on `lsp-mode` by delegating the setup
that was originally done to a function, and scoping the `require` to
that function.

I should *probably* test this on a clean Emacs instance, but at the time
of authoring this commit, it's late as hell, and I'm tired. - Dr. Bluefall